Initial commit of preliminary Bitmap storage layer.
Pieces are as follows:
Bitset - a compressed bitmap implemented with the RoaringBitmap library
Bitmap - a bitmap representation that allows syncronization with files
BitmapFile - provides functions to write and read the file to the bitmap
BitmapFileManager - provides a cache for bitmaps, and creating and loading
BitmapIndex - an entire index representation with access to bitmaps for all values

Added preliminary tests for BitmapFile, Bitmap, and BitmapFileManager. Tests ensure
that bitmaps can be written to disk and read with integrity.